+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-initial release
+Release 1.1.0
+
+- `onDispose` has been added to `StatefulProvider`
+- `BuildContext` is now passed to `valueBuilder` callback